+++ src/main/java/com/bookspot/users/application/UserService.java
@@ -4,7 +4,7 @@ import com.bookspot.users.domain.OAuthProvider;
 import com.bookspot.users.domain.Users;
 import com.bookspot.users.domain.UsersRepository;
 import com.bookspot.users.domain.exception.UserNotFoundException;
-import com.bookspot.users.presentation.UserDetailResponse;
+import com.bookspot.users.presentation.response.UserDetailResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
