+++ src/main/java/com/bookspot/users/presentation/UserAuthController.java
@@ -2,6 +2,8 @@ package com.bookspot.users.presentation;
 
 import com.bookspot.users.application.UserTokenService;
 import com.bookspot.users.domain.OAuthProvider;
+import com.bookspot.users.presentation.request.LoginRequest;
+import com.bookspot.users.presentation.response.UserTokenResponse;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
