+++ src/main/java/com/bookspot/users/presentation/response/UserDetailResponse.java
@@ -1,4 +1,4 @@
-package com.bookspot.users.presentation;
+package com.bookspot.users.presentation.response;
 
 import java.time.LocalDateTime;
